test: split query editor tabs e2e specs
